Makes board row inputs uncontrolled
Using local state to control the input value means that when other
clients update the activity description, the changes aren't reflected in
the local state for the input. Making them uncontrolled allows the
default value to be driven directly by the client cache value, which is
updated whenever someone updates an activity description.